move home to components folder
