+++ src/app/modules/home-login/components/home.component.scss
@@ -1,0 +1,8 @@
+.body {
+  background: url('../../../../assets/longExpo_UW.jpg') no-repeat fixed center center;
+  background-size: cover;
+  min-height: 100vh;
+  overflow: auto;
+}
+
+
