+++ src/app/modules/home-login/components/search-pictures-dialog/search-pictures-dialog.component.scss
